Don't convert initial trial payment method IDs to charge IDs
Exchange requires a method ID when creating a transaction. With subscriptions,
we set this to the subscriber ID. For non-trial payments, Stripe will send a webhook
shortly thereafter with a charge ID representing the first payment of the subscription.

However, for trial subscriptions, the customer is never charged, so no charge ID exists.
Previously, we would still convert the subscription ID to a charge ID once we got a payment,
but this charge ID is representing the first time the customer pays Stripe, but the second
payment record in Exchange. This had the side-effect of blocking the second child
transaction from being created in Exchange.

For subscription's with trials, we no longer convert the subscription ID to a charge ID at all,
and instead just leave the subscription ID as the method ID. Stripe does create an invoice
for this 0.00 payment, but it happens instantaneously making it difficult to use as a method
ID in the current way Exchange handles transactions. This should be revisited once the
payment gateway overhaul is completed.